% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmnet.wrapper.R
\name{cv.glmnet.wrapper}
\alias{cv.glmnet.wrapper}
\title{Add together two numbers.}
\usage{
cv.glmnet.wrapper(y, x, kfold = 10, scale.x = F, verbose = T, ...)
}
\arguments{
\item{y}{a vector of outcomes.}

\item{x}{x matrix}

\item{k}{an integer for the number of folds.}
}
\value{
Elastic Net Results
}
\description{
Add together two numbers.
}
\examples{
cv.glmnet.wrapper()
}
\author{
Olivier Bakker, Raul Aguirre, Matthijs Knigge, Twan Spenkelink
}
