% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.bar.plot.R
\name{glmnet.bar.plot}
\alias{glmnet.bar.plot}
\title{plot elastic net}
\usage{
glmnet.bar.plot(identifier, group, before, after, title)
}
\arguments{
\item{identifier}{vector character of phenotypes}

\item{group}{vector character of grouping}

\item{title}{character title of plot}

\item{gene.amount.before}{numeric vector containing average genes before per imputation}

\item{gene.amount.after}{numeric vector containing average genes after per imputation}
}
\value{
ggplot object
}
\description{
plot elastic net
}
\examples{
glmnet.bar.plot()
}
\author{
Matthijs Knigge
}
\keyword{plot}
