% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.box.plot.R
\name{glmnet.box.plot}
\alias{glmnet.box.plot}
\title{box plot elastic net}
\usage{
glmnet.box.plot(coef, group, identifier)
}
\arguments{
\item{coef}{vector float, correlation coefficients}

\item{group}{vector character of grouping}

\item{identifier}{vector character of phenotypes}
}
\value{
ggplot object
}
\description{
box plot elastic net
}
\examples{
glmnet.box.plot()
}
\author{
Matthijs Knigge
}
\keyword{plot}
