% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.cor.R
\name{glmnet.cor}
\alias{glmnet.cor}
\title{calculate correlation matrix}
\usage{
glmnet.cor(data.x, data.y, write = FALSE, path = NULL, melt = FALSE,
  method = c("spearman"))
}
\arguments{
\item{data.x}{matrix n*m, must have row, and colnames.}

\item{data.y}{matrix n*m, must have row, and colnames.}

\item{write}{boolean, default = FALSE. Set TRUE for storing matrix.}

\item{path}{character, default = NULL. Path to save matrix.}

\item{melt}{boolean, default = FALSE. Set TRUE, if need to return or write melted version of correlation.}
}
\value{
correlation matrix between y and x
}
\description{
calculate correlation matrix
}
\examples{
glmnet.cor()
}
\author{
Matthijs Knigge
}
\keyword{cor}
\keyword{correlation}
