% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.create.mean.labels.R
\name{glmnet.create.mean.labels}
\alias{glmnet.create.mean.labels}
\title{create mean labels}
\usage{
glmnet.create.mean.labels(identifier, group, coef)
}
\arguments{
\item{identifier}{no comment yet}

\item{group}{no comment yet}

\item{coef}{no comment yet}
}
\value{
ggplot object
}
\description{
create mean labels
}
\examples{
glmnet.create.mean.labels()
}
\author{
Matthijs Knigge
}
\keyword{labels}
