% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.density.plot.R
\name{glmnet.density.plot}
\alias{glmnet.density.plot}
\title{density plot elastic net}
\usage{
glmnet.density.plot(coef, group, identifier, title)
}
\arguments{
\item{coef}{vector float, correlation coefficients}

\item{group}{vector character of grouping}

\item{identifier}{vector character of phenotypes}

\item{title}{character title of plot}
}
\value{
ggplot object
}
\description{
density plot elastic net
}
\examples{
glmnet.density.plot()
}
\author{
Matthijs Knigge
}
\keyword{plot}
