% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.grid.plot.R
\name{glmnet.grid.plot}
\alias{glmnet.grid.plot}
\title{plot elastic net}
\usage{
glmnet.grid.plot(p.1, p.2, p.3, p.4, p.5, x.lab, title, save = FALSE,
  return = FALSE, path, width = 10, height = 10)
}
\arguments{
\item{p.1}{ggplot object}

\item{p.2}{ggplot object}

\item{p.3}{ggplot object}

\item{x.lab}{vector character names of x-axis}

\item{title}{title of plot character}

\item{save}{boolean, save plot. default FASLE}

\item{return}{boolean, return instead of save. Default FALSE}

\item{path}{character to path to save, must include name of plot.}

\item{width}{numeric width of plot in cm}

\item{height}{numeric height of plot in cm}
}
\value{
ggplot object
}
\description{
plot elastic net
}
\examples{
glmnet.grid.plot()
}
\author{
Matthijs Knigge
}
\keyword{plot}
