% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.inverse.rank.R
\name{glmnet.inverse.rank}
\alias{glmnet.inverse.rank}
\title{Inverse Rank}
\usage{
glmnet.inverse.rank(phenotype)
}
\arguments{
\item{phenotype}{numeric vector}
}
\value{
forces normal distribution of vector
}
\description{
Inverse Rank
}
\examples{
glmnet.inverse.rank()
}
\author{
Matthijs Knigge
}
\keyword{plot}
