% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.plot.R
\name{glmnet.plot}
\alias{glmnet.plot}
\title{plot glmnet results}
\usage{
glmnet.plot(glmnet.object)
}
\value{
plot object
}
\description{
plot glmnet results
}
\examples{
simple.linear.regression(x = c(1,2,3,4,5,6), y = c(1,2,3,4,5,6))
}
\author{
Matthijs Knigge
}
