% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.scatter.plot.R
\name{glmnet.scatter.plot}
\alias{glmnet.scatter.plot}
\title{Scatter plots}
\usage{
glmnet.scatter.plot(x, y, title, sub.title)
}
\arguments{
\item{x}{numeric vector. Mean correlation of interaction between immune phenotypes and genes}

\item{y}{numeric vector. Mean correlation of interaction between immune phenotypes and genes + genetics}

\item{title}{character. Title of plot}

\item{sub.itle}{character. Title of the plot}
}
\value{

}
\description{
Scatter plots
}
\examples{
glmnet.scatter.plot()
}
\author{
Matthijs Knigge
}
\keyword{linear}
\keyword{regression}
