% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.TMM.normalize.R
\name{glmnet.TMM.normalize}
\alias{glmnet.TMM.normalize}
\title{box plot elastic net}
\usage{
glmnet.TMM.normalize(raw.cell.counts)
}
\arguments{
\item{raw.cell.counts}{character absolute path to raw read counts}

\item{output.file}{character name of the outputfile}
}
\description{
box plot elastic net
}
\examples{
glmnet.TMM.normalize()

}
\author{
Olivier Bakker, Matthijs Knigge
}
\keyword{plot}
