% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.wrapper.R
\name{glmnet.wrapper}
\alias{glmnet.wrapper}
\title{Wrapper for Elastic Net}
\usage{
glmnet.wrapper(y, x, newx = NULL, newy = NULL, alpha.runs = NULL,
  parallel = T, return.type = "custom", method = "spearman",
  use = "pairwise.complete.obs", remove.outliers = F)
}
\arguments{
\item{y}{a vector of outcomes.}

\item{x}{matrix}
}
\value{
Elastic Net Results
}
\description{
Wrapper for Elastic Net
}
\examples{
glmnet.wrapper()
}
\author{
Olivier Bakker, Raul Aguirre, Matthijs Knigge, Twan Spenkelink
}
