% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.coef.R
\name{predict.coef}
\alias{predict.coef}
\title{Predict immune ohenotypes}
\usage{
\method{predict}{coef}(x, coef)
}
\arguments{
\item{x}{matrix n*m, must have row, and colnames.}

\item{coef}{matrix n*m, must have row, and colnames.}
}
\value{
prediceted immune phenotypes
}
\description{
Predict immune ohenotypes
}
\examples{
predict.coef()
}
\author{
Matthijs Knigge
}
\keyword{predict}
