% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.create.mean.labels.R
\name{glmnet.create.mean.labels}
\alias{glmnet.create.mean.labels}
\title{create mean labels}
\usage{
glmnet.create.mean.labels(data)
}
\arguments{
\item{data}{data.frame containing columns coef, group and identifier}
}
\value{
ggplot object
}
\description{
create mean labels
}
\examples{
glmnet.create.mean.labels()
}
\author{
Matthijs Knigge
}
\keyword{labels}
