% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.parse.results.R
\name{glmnet.parse.results}
\alias{glmnet.parse.results}
\title{Parse elastic net results}
\usage{
glmnet.parse.results(dir, folds, data.y, merge = FALSE, data.c, melt = F)
}
\arguments{
\item{dir}{character, absolute path to root folder of results}

\item{folds}{numeric, amount of folds used}

\item{data.y}{matrix n*m, must have row, and colnames. Phenotype data}

\item{merge}{boolean, add group column to frame}

\item{melt}{boolean, default FALSE. melt output}

\item{by}{matrix 2*n, character. containing identifer and group column}
}
\value{
results
}
\description{
Parse elastic net results
}
\examples{
glmnet.parse.results()
}
\author{
Matthijs Knigge
}
\keyword{plot}
