% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.plot.R
\name{glmnet.plot}
\alias{glmnet.plot}
\title{plot elastic net}
\usage{
glmnet.plot(glmnet.data.frame, ylab, xlab, title, interval, filename)
}
\value{
ggplot object
}
\description{
plot elastic net
}
\examples{
glmnet.plot()
}
\author{
Matthijs Knigge
}
\keyword{plot}
