% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.plot.R
\name{glmnet.plot}
\alias{glmnet.plot}
\title{plot elastic net}
\usage{
glmnet.plot(obs, pred, fold, identifier, imp)
}
\arguments{
\item{obs}{vector float, containing predicted values for Y}

\item{pred}{vector float, containing observed values for X}

\item{fold}{vector character, containing fold identifier}

\item{identifier}{character name of phenotype}

\item{imp}{numeric, impuation identifier}
}
\value{
ggplot object
}
\description{
plot elastic net
}
\examples{
glmnet.plot()
}
\author{
Matthijs Knigge
}
\keyword{plot}
